requirements.txt
beautifulsoup4==4.12.2
selenium==4.15.2
requests==2.31.0
newspaper3k==0.2.8
python-dateutil==2.8.2
gspread==5.12.0
google-auth==2.23.4
weasyprint==65.1
trafilatura==1.6.1
readability-lxml==0.8.1
selenium-stealth==1.0.6 
goose3==3.1.19